refactor(read-models): move Wishlist reads to query handlers (PR6b-5)
Final sub-PR of the 6b read-model arc. WishlistViewModel's three direct
DbContext reads move into BookTracker.Application query handlers; the VM
now takes only IBookLookupService + IDispatcher + ILogger (IDbContextFactory
dropped).

New handlers (verbatim relocations — logic preserved exactly, scale
optimisations deferred to TD-17 at arc close):
- GetWishlist        <- LoadWishlistAsync (priority-desc, title list)
- GetWishlistDuplicate <- FindDuplicateMatchesAsync (owned/wishlisted badge)
- GetSeriesGaps      <- LoadSeriesGapsAsync (slot-occupancy math + open series)

DTOs WishlistRow / SeriesGap / OwnedSeriesBook / OpenSeries move to
BookTracker.Application.Wishlist; WishlistCandidate + QuickAddInput stay on
the VM (search-only, never relocated). Index.razor gains the @using and
references the moved types unqualified.

The gnarly series-gaps math keeps its existing regression coverage: the
three LoadSeriesGapsAsync_* VM tests (incl. the floored-interquel guard)
now exercise GetSeriesGapsHandler end-to-end via TestDispatcher.For(factory),
so no duplicate handler-level test is added.
